modified create.sql - changed the data type of posts.user_id to bigint to account for more users

+++ ProjectSourceCode/src/init_data/create.sql
@@ -9,7 +9,7 @@ CREATE TABLE posts (
     id SERIAL PRIMARY KEY,
     img VARCHAR(1000),
     text VARCHAR(500),
-    user_id INT
+    user_id BIGINT
 );
 
 ALTER TABLE posts
